Rename Profile Detail spec file
Fix hasPhoto usage on detail page
